Adds celery upstart files to ansible deployment
We need celery and celerybeat to be running on our box. The other thing
that we ought to enable in our next commit is setting up JUST a celery
box, where the broker variable gets set to the primary web server. That
would enable spinning up multipe task queue servers.