Adds a homepage app with brigades, initiatives and partners
This allows us to keep track of our initiatives and partners in the
database. It also starts mocking in the per-brigade feature of the app.
We'll have to go backwards through the codebase and make everything
depend on a brigade to some extent or another. But that will be later.