Updates views for proper homepage and dashboard views
Before I was hacking it around with a BuzzList view rather than proper
homepage and dashboard views. Now we've got them. The only thing left
for a 2.0.0 release now is to stub in the events at the bottom of the
homepage and add some templates for tech lists and topic lists.